🛠[프로젝트_내 프로젝트] : 내 프로젝트 컨테이너 생성
